Orbit targets to create stability
Created class for bodies in used in the simulation to have "Orbit
targets" the class has a canOrbit() function that decides if the body in
question can orbit the other, the initial body can orbit the other if it
has no target. the other body is the target or they are close together.